refactor(api-file-manager): extract shared upload types to base package
Move FileData, PresignedPostPayloadData, UploadPayloadResponse,
MultiPartUploadFilePart, and CreateMultiPartUploadResult from provider
packages into packages/api-file-manager/src/features/upload/types.ts.
Provider packages now re-export from the base package and keep only
their provider-specific types.
