+++ packages/api-file-manager-s3/src/types.ts
@@ -1,22 +1,9 @@
 import type { PresignedPost } from "@webiny/aws-sdk/client-s3/index.js";
+import type { FileData } from "@webiny/api-file-manager/features/upload/index.js";
 import "@webiny/background-tasks/api/features/TaskController/augmentation.js";
 
-export interface PresignedPostPayloadData {
-    name: string;
-    type: string;
-    size: number;
-    id?: string;
-    key?: string;
-    keyPrefix?: string;
-}
-
-export interface FileData {
-    id: string;
-    key: string;
-    name: string;
-    size: number;
-    type: string;
-}
+export type { PresignedPostPayloadData } from "@webiny/api-file-manager/features/upload/index.js";
+export type { FileData } from "@webiny/api-file-manager/features/upload/index.js";
 
 export interface PresignedPostPayloadDataResponse {
     data: PresignedPost;
